Make foreign function handling selectable, but with common runtime
This uses the stack-frame-less runtime model proposed by Nick
But there are now two different scripts for generating the wrappers,
one ("forward") that conforms to the previous model, and another
("reverse") that conforms to Nick's model.

Given some function foo that we intend to wrap.

"forward" means "foo" is the name of the wrapper, and the
application is unmodified.  FPVM then can invoke fpvm_foo
or go directly via the __orig pointer to invok,e the
underlying function

"reverse" means "foo$fpvm" is the name of the wrapper.
The assumption is that the application has been modified
using (for example, a lief script) so that every reference
to foo is now a refrence to foo$fpvm.   FPVM can then
invoke foo directly.