Executable regions cannot contain GC-able values
This is particularly important so we do not scan all of the .so's that
were mapped in during init. On ARM & RISC-V, we need to do the
debugger trick of making these executable memory regions writable so
we can drop a breakpoint on them. However, since FPVM should never use
a breakpoint, this should be a moot point.

Linux enforces W^X (write XOR execute) on memory regions by default.
We assume (fairly) that code pages that are executable do not contain
NaNs that FPVM created with its GC.