Add fptrapall masking to profiler/fpvm_magic.c
Need to mask/unmask the CR0.ts bit on entry to FPVM through magic traps.
Currently the masking/unmasking isn't nestable, (which would slow things
down) but we should add some state to allow that. (We could track the
state of the TS bit in software to avoid asking the kernel whether it is
set or not).