Added fpvm_op_to_string (Using an xmacro...)
I got sick of tracking down the association between opcodes numerically
and the actual enum value from error messgaes, so I quickly changed the
definition of FPVM_OP_* to be in an xmacro, which let me generate a
simple "to_string" function for them.

So now all error messages in the emulator about an opcode can just print
out the actual "stringified" name.