Add I2F conversions to op_map but ignore them
CR0.ts causes us to actually see I2F conversions faulting, but FPVM just
seems to fall over when it sees one, so this is set up to automatically
fail in the emulator when an I2F instruction is encountered. Which means
we can then single step over the conversion instruction.