Add FPVM_TEENY_ROUND_TOO_SMALLS_AWAY_FROM_ZERO feature
If this environment variable is set to "y", then the
teeny altmath will round away from zero when encoding
a double that would normally round to zero.

More specifically, if the double is either a norm
or a subnorm with a nonzero mantissa, and the result
would underflow the teeny to zero normally, it will
instead underflow it to the smallest possible magnitude
nonzero teeny subnorm.